+++ handbook/src/pages/index.js
@@ -163,7 +163,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#1fd898" }}>222,957</div>
+            <div style={{ color: "#1fd898" }}>224,936</div>
             <span className={isDarkTheme ? " dark" : ""}>Downloads</span>
           </div>
         </div>